playbooks/roles/client/files/client/requirements.txt
pulsar-client==2.6.1
grpcio<1.26.0